requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.0.1
pdf2image==1.16.3
requests==2.31.0
Jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0